Reorganize requirements
Waves 2 and 3 were not ordered in terms of their implementation
difficulty, but instead because scoring point seemed more central to the
gameplay and thus was implemented earlier.

However, this reorganization also seems to follow a logical gameplay
progression (spotting items leading to scoring points), while also
having a steadier progression of complexity.